fix(cat): move visual editor types out of vercelignored fixture
Production CAT visual editor components imported types from
cat-visual-editor.fixture.ts, which .vercelignore excludes, so Vercel
typecheck failed with missing-module errors after #1781.
